Set up project base
Add a the base structure to realize the project. It is not functional right now at it do not exposes the API (routes are kinda empty) but it add a base for the database as well as migration.

It is using:
- Flask
- Flask-Migrate (with alembic)
- Flask-SQLAlchemy (with SQLAlchemy)
- SQLAlchemy-Utils: use helpers for some fields such as email addresses

Next steps would be to create some unit tests to assert the integrity of the database.